Move rendering of ticket cards into its own view
It’s a bit more readable, but also solves a nasty bug.

When the answers would update (after loading of the view), the view would rerender. When this would happen during the accessing of the pasteboard, that dialog would be auto-dismissed and the pasteboard access would fail.

Now the updating of the answers only triggers a rerender of the cards, and does not affect the pasteboard access at all.